# WPF_CONTACT_V1/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WEB_API_CONTACT",
    "version": "1.0"
  },
  "paths": {
    "/CONTACTS/GetDate": {
      "get": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "GetDate",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/CONTACTS/GetAllContacts": {
      "get": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "GetAllContacts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_CONTACT"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_CONTACT"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_CONTACT"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/CONTACTS/AddContact": {
      "post": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "AddContact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_CONTACT"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CONTACT"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CONTACT"
                }
              }
            }
          }
        }
      }
    },
    "/CONTACTS/Save": {
      "get": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "Save",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CONTACTS/DeleteContact": {
      "delete": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "DeleteContact",
        "parameters": [
          {
            "name": "P_ID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CONTACTS/UpdateContact": {
      "put": {
        "tags": [
          "CONTACT"
        ],
        "operationId": "UpdateContact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_CONTACT"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "C_CONTACT": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string",
            "nullable": true
          },
          "prenom": {
            "type": "string",
            "nullable": true
          },
          "mail": {
            "type": "string",
            "nullable": true
          },
          "num_Tel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}